+++ proton-mail-uniffi/src/mail/user_session/events.rs
@@ -1,5 +1,5 @@
 use crate::mail::MailUserSession;
-use crate::uniffi_async;
+use crate::{uniffi_async, EventCallback};
 use anyhow::anyhow;
 use proton_api_core::service::ApiServiceError;
 use proton_event_loop::subscriber::SubscriberError;
@@ -11,14 +11,26 @@ impl MailUserSession {
     /// Poll Event loop and apply events.
     ///
     /// *NOTE*: do not call this function concurrently.
+    ///
+    /// ## Parameters
+    ///
+    /// callback: EventCallback interface, to run in a background thread watching for updates.
+    ///
     #[allow(clippy::unused_async)]
-    pub async fn poll_events(&self) -> Result<(), EventLoopError> {
+    pub async fn poll_events(
+        &self,
+        callback: Box<dyn EventCallback>,
+    ) -> Result<(), EventLoopError> {
         let ctx = self.ctx.clone();
-        uniffi_async(async move {
+        uniffi_async::<(), EventLoopError, _>(async move {
             ctx.poll_event_loop().await?;
             Ok(())
         })
-        .await
+        .await?;
+
+        callback.on_update();
+
+        Ok(())
     }
 }